requirements.txt
fastapi
mangum
pytest
aws-lambda-powertools
jmespath
boto3
cdk
requests
pyyaml